move each page to proper module and folder
