+++ spec/support/pages/sections/menu.rb
@@ -1,0 +1,5 @@
+module Sections
+  class Menu < SitePrism::Section
+    element :admin_panel_link, "a[href='/admin/users']"
+  end
+end
